+++ scripts/end-round.py
@@ -3,7 +3,8 @@ import sys
 
 from server.models import Election
 from server.database import db_session
-from server.api.rounds import get_current_round, is_round_complete, end_round
+from server.api.shared import get_current_round, is_round_complete
+from server.api.rounds import end_round
 
 if __name__ == "__main__":
     if len(sys.argv) != 2:
