+++ server/api/rounds.py
@@ -1,36 +1,50 @@
-import random
 import uuid
-from collections import defaultdict
 from typing import (
-    Optional,
     List,
     Tuple,
     Dict,
-    TypedDict,
 )
 from datetime import datetime
 from flask import jsonify, request
 from jsonschema import validate
 from werkzeug.exceptions import BadRequest, Conflict
-from sqlalchemy import and_, func, not_, literal
-from sqlalchemy.orm import joinedload
+from sqlalchemy import func, not_
+
 
 from . import api
 from ..database import db_session
 from ..models import *  # pylint: disable=wildcard-import
+from .shared import (
+    BallotDraw,
+    SampleSize,
+    active_targeted_contests,
+    batch_tallies,
+    compute_sample_ballots,
+    compute_sample_batches,
+    contest_results_by_round,
+    cvrs_for_contest,
+    full_hand_tally_sizes,
+    get_current_round,
+    get_previous_round,
+    hybrid_contest_strata,
+    is_full_hand_tally,
+    needs_full_hand_tally,
+    round_sizes,
+    sampled_ballot_interpretations_to_cvrs,
+    sampled_batch_results,
+    sampled_batches_by_ticket_number,
+    samples_not_found_by_round,
+)
 from ..auth import restrict_access, UserType
 from ..util.isoformat import isoformat
 from ..util.collections import group_by
 from ..audit_math import (
-    sampler,
     ballot_polling,
-    ballot_polling_types,
     macro,
     supersimple,
     sampler_contest,
     suite,
 )
-from .cvrs import hybrid_contest_choice_vote_counts, cvr_contests_metadata
 from .ballot_manifest import hybrid_contest_total_ballots
 from ..worker.tasks import (
     background_task,
@@ -44,31 +58,6 @@ from ..activity_log import (
     EndRound,
 )
 from ..feature_flags import is_enabled_sample_extra_batches_by_counting_group
-
-
-def get_current_round(election: Election) -> Optional[Round]:
-    if len(list(election.rounds)) == 0:
-        return None
-    return max(election.rounds, key=lambda r: r.round_num)
-
-
-def get_previous_round(election: Election, round: Round) -> Optional[Round]:
-    if round.round_num == 1:
-        return None
-    return next(r for r in election.rounds if r.round_num == round.round_num - 1)
-
-
-# Returns a list of targeted contests that are still being audited (i.e. haven't
-# yet met the risk limit)
-def active_targeted_contests(election: Election) -> List[Contest]:
-    targeted_contests = Contest.query.filter_by(
-        election_id=election.id, is_targeted=True
-    )
-    return list(
-        targeted_contests.all()
-        if len(list(election.rounds)) == 0
-        else targeted_contests.join(RoundContest).filter_by(is_complete=False).all()
-    )
 
 
 def is_audit_complete(round: Round):
@@ -177,368 +166,6 @@ def count_audited_votes(election: Election, round: Round):
             db_session.add(result)
 
 
-def contest_results_by_round(
-    contest: Contest,
-) -> Optional[ballot_polling_types.BALLOT_POLLING_SAMPLE_RESULTS]:
-    results_by_round: ballot_polling_types.BALLOT_POLLING_SAMPLE_RESULTS = defaultdict(
-        lambda: defaultdict(int)
-    )
-    for result in contest.results:
-        results_by_round[result.round_id][result.contest_choice_id] = result.result
-    return results_by_round if len(results_by_round) > 0 else None
-
-
-def samples_not_found_by_round(contest: Contest) -> Dict[str, int]:
-    if contest.is_targeted:
-        return dict(
-            SampledBallotDraw.query.filter_by(contest_id=contest.id)
-            .join(SampledBallot)
-            .filter_by(status=BallotStatus.NOT_FOUND)
-            .group_by(SampledBallotDraw.round_id)
-            .values(SampledBallotDraw.round_id, func.count())
-        )
-    else:
-        return dict(
-            SampledBallot.query.filter_by(status=BallotStatus.NOT_FOUND)
-            .join(Batch)
-            .join(Jurisdiction)
-            .join(Jurisdiction.contests)
-            .filter_by(id=contest.id)
-            .join(SampledBallot.draws)
-            .group_by(SampledBallotDraw.round_id)
-            .values(SampledBallotDraw.round_id, func.count(SampledBallot.id.distinct()))
-        )
-
-
-# { batch_key: { contest_id: { choice_id: votes }}}
-BatchTallies = Dict[macro.BatchKey, macro.BatchResults]
-
-
-def batch_tallies(election: Election) -> BatchTallies:
-    # We only support one contest for batch audits
-    assert len(list(election.contests)) == 1
-    contest = list(election.contests)[0]
-
-    # Key each batch by jurisdiction name and batch name since batch names
-    # are only guaranteed unique within a jurisdiction
-    return {
-        (jurisdiction.name, batch_name): tally
-        for jurisdiction in contest.jurisdictions
-        for batch_name, tally in jurisdiction.batch_tallies.items()  # type: ignore
-    }
-
-
-def sampled_batch_results(election: Election) -> BatchTallies:
-    results_by_batch_and_choice = (
-        Batch.query.filter(
-            Batch.id.in_(
-                Batch.query.join(Jurisdiction)
-                .filter_by(election_id=election.id)
-                .join(SampledBatchDraw)
-                # Special case: don't include extra sampled batches
-                .filter(SampledBatchDraw.ticket_number != EXTRA_TICKET_NUMBER)
-                .values(Batch.id)
-            )
-        )
-        .join(Jurisdiction)
-        .join(Jurisdiction.contests)
-        .join(ContestChoice)
-        .outerjoin(BatchResultTallySheet)
-        .outerjoin(
-            BatchResult,
-            and_(
-                BatchResult.tally_sheet_id == BatchResultTallySheet.id,
-                BatchResult.contest_choice_id == ContestChoice.id,
-            ),
-        )
-        .group_by(Jurisdiction.name, Batch.name, ContestChoice.id)
-        .values(
-            Jurisdiction.name,
-            Batch.name,
-            ContestChoice.id,
-            func.sum(func.coalesce(BatchResult.result, 0)),
-        )
-    )
-    results_by_batch = group_by(
-        results_by_batch_and_choice,
-        key=lambda result: (result[0], result[1]),  # (jurisdiction_name, batch_name)
-    )
-    # We only support one contest for batch audits
-    assert len(list(election.contests)) == 1
-    contest_id = list(election.contests)[0].id
-    return {
-        batch_key: {
-            contest_id: {
-                choice_id: result for (_, _, choice_id, result) in batch_results
-            }
-        }
-        for batch_key, batch_results in results_by_batch.items()
-    }
-
-
-def sampled_batches_by_ticket_number(election: Election) -> Dict[str, sampler.BatchKey]:
-    batches_by_ticket_number = (
-        SampledBatchDraw.query.join(Batch)
-        .join(Jurisdiction)
-        .filter_by(election_id=election.id)
-        # Special case: don't include extra sampled batches
-        .filter(SampledBatchDraw.ticket_number != EXTRA_TICKET_NUMBER)
-        .order_by(SampledBatchDraw.ticket_number)
-        .values(SampledBatchDraw.ticket_number, Jurisdiction.name, Batch.name)
-    )
-    return {
-        ticket_number: (jurisdiction_name, batch_name)
-        for ticket_number, jurisdiction_name, batch_name in batches_by_ticket_number
-    }
-
-
-def round_sizes(contest: Contest) -> ballot_polling_types.BALLOT_POLLING_ROUND_SIZES:
-    # For targeted contests, return the number of ballots sampled for that contest
-    if contest.is_targeted:
-        results = (
-            Round.query.join(SampledBallotDraw)
-            .filter_by(contest_id=contest.id)
-            .group_by(Round.id, Round.round_num)
-            .values(
-                Round.round_num, Round.id, func.count(SampledBallotDraw.ticket_number)
-            )
-        )
-        return {
-            round_num: ballot_polling_types.RoundInfo(round_id, count)
-            for round_num, round_id, count in results
-        }
-    # For opportunistic contests, return the number of sampled ballots in
-    # jurisdictions in that contest's universe
-    else:
-        contest_jurisdiction_ballots = (
-            SampledBallot.query.join(Batch)
-            .join(Jurisdiction)
-            .join(Jurisdiction.contests)
-            .filter_by(id=contest.id)
-            .with_entities(SampledBallot.id)
-            .subquery()
-        )
-        results = (
-            Round.query.join(SampledBallotDraw)
-            .join(SampledBallot)
-            .filter(SampledBallot.id.in_(contest_jurisdiction_ballots))
-            .group_by(Round.id, Round.round_num)
-            .values(Round.round_num, Round.id, func.count(SampledBallot.id.distinct()))
-        )
-        return {
-            round_num: ballot_polling_types.RoundInfo(round_id, count)
-            for round_num, round_id, count in results
-        }
-
-
-def cvrs_for_contest(contest: Contest) -> sampler_contest.CVRS:
-    cvrs: sampler_contest.CVRS = {}
-
-    ballot_interpretations = (
-        CvrBallot.query.join(Batch)
-        .join(Jurisdiction)
-        .join(Jurisdiction.contests)
-        .filter_by(id=contest.id)
-        .join(
-            SampledBallot,
-            and_(
-                CvrBallot.batch_id == SampledBallot.batch_id,
-                CvrBallot.ballot_position == SampledBallot.ballot_position,
-            ),
-        )
-        .values(Jurisdiction.id, SampledBallot.id, CvrBallot.interpretations)
-    )
-
-    metadata_by_jurisdictions = {
-        jurisdiction.id: cvr_contests_metadata(jurisdiction)
-        for jurisdiction in contest.jurisdictions
-    }
-
-    for (jurisdiction_id, ballot_key, interpretations_str) in ballot_interpretations:
-        metadata = metadata_by_jurisdictions[jurisdiction_id]
-        assert metadata is not None
-        choices_metadata = metadata[contest.name]["choices"]
-
-        # interpretations is the raw CVR string: 1,0,0,1,0,1,0. We need to
-        # pick out the interpretation for each contest choice. We saved the
-        # column index for each choice when we parsed the CVR.
-        interpretations = interpretations_str.split(",")
-        choice_interpretations = {
-            choice_name: interpretations[choice_metadata["column"]]
-            for choice_name, choice_metadata in choices_metadata.items()
-        }
-
-        # If the interpretations are empty, it means the contest wasn't
-        # on the ballot, so we should skip this contest entirely for
-        # this ballot.
-        if all(
-            interpretation == "" for interpretation in choice_interpretations.values()
-        ):
-            cvrs[ballot_key] = {}
-        else:
-            # Parse each choice's interpretation. We use the main list of
-            # contest choices since each jurisdiction's CVR may only record
-            # a subset of the choices (e.g. in ES&S/Hart). If there's a
-            # choice we don't have a CVR interpretation for, we can assume
-            # it didn't get voted for and set its interpretation to 0.
-            cvrs[ballot_key] = {
-                contest.id: {
-                    choice.id: choice_interpretations.get(choice.name, "0")
-                    for choice in contest.choices
-                }
-            }
-
-    return cvrs
-
-
-def sampled_ballot_interpretations_to_cvrs(
-    contest: Contest,
-) -> sampler_contest.SAMPLECVRS:
-    ballots_query = SampledBallot.query.join(Batch)
-
-    # In hybrid audits, only count CVR ballots
-    if contest.election.audit_type == AuditType.HYBRID:
-        ballots_query = ballots_query.filter(
-            Batch.has_cvrs == True  # pylint: disable=singleton-comparison
-        )
-
-    # For targeted contests, count the number of times the ballot was sampled
-    if contest.is_targeted:
-        ballots_query = (
-            ballots_query.join(SampledBallotDraw)
-            .filter_by(contest_id=contest.id)
-            .group_by(SampledBallot.id)
-            .with_entities(SampledBallot, func.count(SampledBallotDraw.ticket_number))
-        )
-    # For opportunistic contests, we say each ballot was only sampled once
-    else:
-        ballots_query = (
-            ballots_query.join(Jurisdiction)
-            .join(Jurisdiction.contests)
-            .filter_by(id=contest.id)
-            .with_entities(SampledBallot, literal(1))
-        )
-
-    ballots = ballots_query.options(
-        joinedload(SampledBallot.interpretations)
-        .joinedload(BallotInterpretation.selected_choices)
-        .load_only(ContestChoice.id)
-    ).all()
-
-    # The CVR we build should have a 1 for each choice that got voted for,
-    # and a 0 otherwise. There are a couple special cases:
-    # - Contest wasn't on the ballot - CVR should be an empty object
-    # - Audit board couldn't find the ballot - CVR should be None
-    cvrs: sampler_contest.SAMPLECVRS = {}
-    for ballot, times_sampled in ballots:
-        if ballot.status == BallotStatus.NOT_FOUND:
-            cvrs[ballot.id] = {"times_sampled": times_sampled, "cvr": None}
-
-        elif ballot.status == BallotStatus.AUDITED:
-            interpretation = next(
-                (
-                    interpretation
-                    for interpretation in ballot.interpretations
-                    if interpretation.contest_id == contest.id
-                ),
-                None,
-            )
-            if (
-                interpretation is None  # Legacy case for contest not on ballot
-                or interpretation.interpretation == Interpretation.CONTEST_NOT_ON_BALLOT
-            ):
-                ballot_cvr = {}
-            elif interpretation.interpretation == Interpretation.BLANK:
-                ballot_cvr = {
-                    contest.id: {choice.id: "0" for choice in contest.choices}
-                }
-            elif interpretation.interpretation == Interpretation.VOTE:
-                ballot_cvr = {
-                    contest.id: {
-                        choice.id: (
-                            "1"
-                            if any(
-                                selected_choice.id == choice.id
-                                for selected_choice in interpretation.selected_choices
-                            )
-                            else "0"
-                        )
-                        for choice in contest.choices
-                    }
-                }
-            else:
-                raise Exception(
-                    f"Unexpected interpretation type: {interpretation.interpretation}"
-                )  # pragma: no cover
-
-            cvrs[ballot.id] = {"times_sampled": times_sampled, "cvr": ballot_cvr}
-
-    return cvrs
-
-
-def hybrid_contest_strata(
-    contest: Contest,
-) -> Tuple[suite.BallotPollingStratum, suite.BallotComparisonStratum]:
-    total_ballots = hybrid_contest_total_ballots(contest)
-    vote_counts = hybrid_contest_choice_vote_counts(contest)
-    assert vote_counts
-    non_cvr_vote_counts = {
-        choice_id: vote_count.non_cvr for choice_id, vote_count in vote_counts.items()
-    }
-    cvr_vote_counts = {
-        choice_id: vote_count.cvr for choice_id, vote_count in vote_counts.items()
-    }
-
-    # For targeted contests, count the number of samples drawn for this
-    # contest so far
-    if contest.is_targeted:
-        num_previous_samples_dict = dict(
-            SampledBallotDraw.query.filter_by(contest_id=contest.id)
-            .join(SampledBallot)
-            .join(Batch)
-            .group_by(Batch.has_cvrs)
-            .values(Batch.has_cvrs, func.count(SampledBallotDraw.ticket_number))
-        )
-    # For opportunistic contests, count the number of ballots in jurisdictions
-    # in this contest's universe that were sampled (for some targeted contest)
-    else:
-        num_previous_samples_dict = dict(
-            SampledBallot.query.join(Batch)
-            .join(Jurisdiction)
-            .join(Jurisdiction.contests)
-            .filter_by(id=contest.id)
-            .group_by(Batch.has_cvrs)
-            .values(Batch.has_cvrs, func.count(SampledBallot.id))
-        )
-
-    non_cvr_previous_samples = num_previous_samples_dict.get(False, 0)
-    cvr_previous_samples = num_previous_samples_dict.get(True, 0)
-
-    # In hybrid audits, we only store round contest results for non-CVR
-    # ballots
-    non_cvr_sample_results = contest_results_by_round(contest) or {}
-    non_cvr_stratum = suite.BallotPollingStratum(
-        total_ballots.non_cvr,
-        non_cvr_vote_counts,
-        non_cvr_sample_results,
-        non_cvr_previous_samples,
-    )
-
-    cvr_reported_results = cvrs_for_contest(contest)
-    # The CVR sample results are filtered to only CVR ballots
-    suite_contest = sampler_contest.from_db_contest(contest)
-    cvr_sample_results = sampled_ballot_interpretations_to_cvrs(contest)
-    cvr_misstatements = suite.misstatements(
-        suite_contest, cvr_reported_results, cvr_sample_results,
-    )
-    # Create a stratum for CVR ballots
-    cvr_stratum = suite.BallotComparisonStratum(
-        total_ballots.cvr, cvr_vote_counts, cvr_misstatements, cvr_previous_samples,
-    )
-
-    return non_cvr_stratum, cvr_stratum
-
-
 def calculate_risk_measurements(election: Election, round: Round):
     assert election.risk_limit is not None
 
@@ -599,412 +226,6 @@ def end_round(election: Election, round: Round):
             is_audit_complete=is_audit_complete(round),
         )
     )
-
-
-def is_round_complete(election: Election, round: Round) -> bool:
-    # For batch audits, check that all jurisdictions with sampled batches this
-    # round have finalized batch results
-    if election.audit_type == AuditType.BATCH_COMPARISON:
-        num_jurisdictions_not_finalized: int = (
-            Jurisdiction.query.filter_by(election_id=election.id)
-            .filter(
-                Jurisdiction.id.in_(
-                    SampledBatchDraw.query.filter_by(round_id=round.id)
-                    .join(Batch)
-                    .with_entities(Batch.jurisdiction_id)
-                    .subquery()
-                )
-            )
-            .filter(
-                Jurisdiction.id.notin_(
-                    BatchResultsFinalized.query.filter_by(round_id=round.id)
-                    .with_entities(BatchResultsFinalized.jurisdiction_id)
-                    .subquery()
-                )
-            )
-            .with_entities(Jurisdiction.id)
-            .count()
-        )
-        return num_jurisdictions_not_finalized == 0
-
-    # For online audits, check that all the audit boards are finished auditing
-    if election.online:
-        num_jurisdictions_without_audit_boards_set_up: int = (
-            # For each jurisdiction...
-            Jurisdiction.query.filter_by(election_id=election.id)
-            # Where there are ballots that haven't been audited...
-            .join(Jurisdiction.batches)
-            .join(Batch.ballots)
-            .filter(SampledBallot.status == BallotStatus.NOT_AUDITED)
-            # And those ballots got sampled this round...
-            .join(SampledBallot.draws)
-            .filter_by(round_id=round.id)
-            # Count the number of audit boards set up.
-            .outerjoin(
-                AuditBoard,
-                and_(
-                    AuditBoard.jurisdiction_id == Jurisdiction.id,
-                    AuditBoard.round_id == round.id,
-                ),
-            )
-            .group_by(Jurisdiction.id)
-            # Finally, count how many jurisdictions have no audit boards set up.
-            .having(func.count(AuditBoard.id) == 0)
-            .count()
-        )
-        num_audit_boards_with_ballots_not_signed_off: int = (
-            AuditBoard.query.filter_by(round_id=round.id, signed_off_at=None)
-            .join(SampledBallot)
-            .group_by(AuditBoard.id)
-            .count()
-        )
-        return (
-            num_jurisdictions_without_audit_boards_set_up == 0
-            and num_audit_boards_with_ballots_not_signed_off == 0
-        )
-
-    # For offline audits, check that we have results recorded for every
-    # jurisdiction that had ballots sampled
-    else:
-        num_jurisdictions_without_results: int
-        # Special case: if we are in a full hand tally, we just check every
-        # jurisdiction in the targeted contest's universe
-        if is_full_hand_tally(round, election):
-            num_jurisdictions_without_results = (
-                Contest.query.filter_by(election_id=election.id, is_targeted=True)
-                .join(Contest.jurisdictions)
-                .outerjoin(
-                    JurisdictionResult,
-                    and_(
-                        JurisdictionResult.jurisdiction_id == Jurisdiction.id,
-                        JurisdictionResult.contest_id == Contest.id,
-                        JurisdictionResult.round_id == round.id,
-                    ),
-                )
-                .group_by(Jurisdiction.id, Contest.id)
-                .having(func.count(JurisdictionResult.result) == 0)
-                .count()
-            )
-        else:
-            num_jurisdictions_without_results = (
-                # For each jurisdiction...
-                Jurisdiction.query.filter_by(election_id=election.id)
-                # Where ballots were sampled...
-                .join(Jurisdiction.batches)
-                .join(Batch.ballots)
-                # And those ballots were sampled this round...
-                .join(SampledBallot.draws)
-                .filter_by(round_id=round.id)
-                .join(SampledBallotDraw.contest)
-                # Count the number of results recorded for each contest.
-                .outerjoin(
-                    JurisdictionResult,
-                    and_(
-                        JurisdictionResult.jurisdiction_id == Jurisdiction.id,
-                        JurisdictionResult.contest_id == Contest.id,
-                        JurisdictionResult.round_id == round.id,
-                    ),
-                )
-                # Finally, count the number of jurisdiction/contest pairs for which
-                # there is no result recorded
-                .group_by(Jurisdiction.id, Contest.id)
-                .having(func.count(JurisdictionResult.result) == 0)
-                .count()
-            )
-        return num_jurisdictions_without_results == 0
-
-
-# Calculates the sample size threshold to trigger a full hand tally in each
-# targeted contest
-def full_hand_tally_sizes(election: Election):
-    contests_query = Contest.query.filter_by(election_id=election.id, is_targeted=True)
-    if election.audit_type == AuditType.BATCH_COMPARISON:
-        return dict(
-            contests_query.join(Contest.jurisdictions)
-            .group_by(Contest.id)
-            .values(
-                Contest.id,
-                func.coalesce(func.sum(Jurisdiction.manifest_num_batches), 0),
-            )
-        )
-    return dict(contests_query.values(Contest.id, Contest.total_ballots_cast))
-
-
-# Returns True if the cumulative sample size up to the specified round for any targeted contest
-# indicates that a full hand tally is needed
-def needs_full_hand_tally(round: Round, election: Election) -> bool:
-    full_hand_tally_size = full_hand_tally_sizes(election)
-    cumulative_sample_sizes = dict(
-        RoundContest.query.join(Round)
-        .filter_by(election_id=round.election_id)
-        .filter(Round.round_num <= round.round_num)
-        .join(RoundContest.contest)
-        .filter_by(is_targeted=True)
-        .group_by(RoundContest.contest_id)
-        .values(
-            RoundContest.contest_id,
-            func.sum(RoundContest.sample_size["size"].as_integer()),
-        )
-    )
-    return any(
-        size >= full_hand_tally_size[contest_id]
-        for contest_id, size in cumulative_sample_sizes.items()
-    )
-
-
-# Returns True if Arlo is in full hand tally mode, which is only triggered
-# in the first round of ballot polling audits when the sample size requires a
-# full hand tally
-def is_full_hand_tally(round: Round, election: Election):
-    return (
-        election.audit_type == AuditType.BALLOT_POLLING
-        and round.round_num == 1
-        and needs_full_hand_tally(round, election)
-    )
-
-
-class SampleSize(TypedDict):
-    size: int
-    key: str
-    prob: Optional[float]
-    sizeCvr: int  # Only in hybrid audits
-    sizeNonCvr: int  # Only in hybrid audits
-
-
-class BallotDraw(TypedDict):
-    batch_id: str
-    ballot_position: int
-    contest_id: str
-    ticket_number: str
-
-
-class BatchDraw(TypedDict):
-    batch_id: str
-    ticket_number: str
-
-
-def compute_sample_batches(
-    election: Election,
-    round_num: int,
-    contest_sample_sizes: List[Tuple[Contest, SampleSize]],
-) -> List[BatchDraw]:
-    # We only support one contest for batch audits
-    assert len(contest_sample_sizes) == 1
-    contest, sample_size = contest_sample_sizes[0]
-
-    # Create a mapping from batch keys used in the sampling back to batch ids
-    batches = (
-        Batch.query.join(Jurisdiction)
-        .filter_by(election_id=election.id)
-        .with_entities(Jurisdiction.name, Batch.name, Batch.id)
-    )
-    batch_key_to_id = {
-        (jurisdiction_name, batch_name): batch_id
-        for jurisdiction_name, batch_name, batch_id in batches
-    }
-
-    previously_sampled_batch_keys: List[sampler.BatchKey] = list(
-        SampledBatchDraw.query.join(Batch)
-        .join(Jurisdiction)
-        .filter_by(election_id=election.id)
-        .with_entities(Jurisdiction.name, Batch.name)
-    )
-
-    sample = sampler.draw_ppeb_sample(
-        str(election.random_seed),
-        sampler_contest.from_db_contest(contest),
-        sample_size["size"],
-        previously_sampled_batch_keys,
-        batch_tallies(election),
-    )
-
-    sample_batches = [
-        BatchDraw(batch_id=batch_key_to_id[batch_key], ticket_number=ticket_number)
-        for ticket_number, batch_key in sample
-    ]
-
-    # Experimental feature
-    # Add extra batches on top of the original sample that will be audited, but
-    # not counted in the final risk measurement.
-    if is_enabled_sample_extra_batches_by_counting_group(election) and round_num == 1:
-        rand = random.Random(str(election.random_seed))
-        for jurisdiction in contest.jurisdictions:
-            batch_ids_with_container_and_num_ballots = (
-                Batch.query.filter_by(jurisdiction_id=jurisdiction.id)
-                .with_entities(Batch.id, Batch.container, Batch.num_ballots)
-                .all()
-            )
-            batch_ids = [
-                batch_id for batch_id, _, _ in batch_ids_with_container_and_num_ballots
-            ]
-            batch_id_to_num_ballots = {
-                batch_id: num_ballots
-                for batch_id, _, num_ballots in batch_ids_with_container_and_num_ballots
-            }
-            # To simplify this experiment, we specify the counting group in the
-            # container column of the ballot manifest
-            bmd_batch_ids = {
-                batch_id
-                for batch_id, container, _ in batch_ids_with_container_and_num_ballots
-                if container
-                in [
-                    "Advanced Voting",
-                    "Advance Voting",
-                    "Election Day",
-                    "Elections Day",
-                ]
-            }
-            hmpb_batch_ids = {
-                batch_id
-                for batch_id, container, _ in batch_ids_with_container_and_num_ballots
-                if container in ["Absentee by Mail", "Provisional"]
-            }
-            sampled_batch_ids = {
-                batch_key_to_id[batch_key]
-                for _, batch_key in sample
-                if batch_key[0] == jurisdiction.name
-            }
-            extra_batch_ids = set()
-            # If we didn't sample any BMD batches, add one to the sample
-            if len(bmd_batch_ids & sampled_batch_ids) == 0 and len(bmd_batch_ids) > 0:
-                extra_bmd_batch_id = rand.choice(list(bmd_batch_ids))
-                extra_batch_ids.add(extra_bmd_batch_id)
-                sample_batches.append(
-                    BatchDraw(
-                        batch_id=extra_bmd_batch_id, ticket_number=EXTRA_TICKET_NUMBER
-                    )
-                )
-            # If we didn't sample any HMPB batches, add one to the sample
-            if len(hmpb_batch_ids & sampled_batch_ids) == 0 and len(hmpb_batch_ids) > 0:
-                extra_hmpb_batch_id = rand.choice(list(hmpb_batch_ids))
-                extra_batch_ids.add(extra_hmpb_batch_id)
-                sample_batches.append(
-                    BatchDraw(
-                        batch_id=extra_hmpb_batch_id, ticket_number=EXTRA_TICKET_NUMBER
-                    )
-                )
-
-            # Continue adding batches until the percentage of jurisdiction ballots selected is at
-            # least 2%
-            min_percentage_of_jurisdiction_ballots_selected = 0.02
-
-            def compute_percentage_of_jurisdiction_ballots_selected(
-                selected_batch_ids, num_jurisdiction_ballots
-            ):
-                num_jurisdiction_ballots_selected = sum(
-                    batch_id_to_num_ballots[batch_id] for batch_id in selected_batch_ids
-                )
-                return num_jurisdiction_ballots_selected / num_jurisdiction_ballots
-
-            while (
-                compute_percentage_of_jurisdiction_ballots_selected(
-                    sampled_batch_ids.union(extra_batch_ids),
-                    jurisdiction.manifest_num_ballots,
-                )
-                < min_percentage_of_jurisdiction_ballots_selected
-            ):
-                remaining_batch_ids = (
-                    set(batch_ids) - sampled_batch_ids - extra_batch_ids
-                )
-                extra_batch_id = rand.choice(list(remaining_batch_ids))
-                extra_batch_ids.add(extra_batch_id)
-                sample_batches.append(
-                    BatchDraw(
-                        batch_id=extra_batch_id, ticket_number=EXTRA_TICKET_NUMBER
-                    )
-                )
-
-    return sample_batches
-
-
-def compute_sample_ballots(
-    election: Election,
-    contest_sample_sizes: List[Tuple[Contest, SampleSize]],
-    # For hybrid audits only, Batch.has_cvrs will be true/false if the batch
-    # contains ballots with CVRs or not (based on the manifest).
-    # filter_has_cvrs will constrain the ballots to sample based on
-    # Batch.has_cvrs. Since Batch.has_cvrs is None for all other audit types,
-    # the default filter is None.
-    filter_has_cvrs: bool = None,
-) -> List[BallotDraw]:
-    participating_jurisdictions = {
-        jurisdiction
-        for (contest, _) in contest_sample_sizes
-        for jurisdiction in contest.jurisdictions
-    }
-
-    # Audits must be deterministic and repeatable for the same real world
-    # inputs. So the sampler expects the same input for the same real world
-    # data. Thus, we use the jurisdiction name and batch keys
-    # (deterministic real world ids) instead of the jurisdiction and batch
-    # ids (non-deterministic uuids that we generate for each audit).
-    batch_id_to_key = {
-        batch.id: (jurisdiction.name, batch.tabulator, batch.name)
-        for jurisdiction in participating_jurisdictions
-        for batch in jurisdiction.batches
-    }
-    batch_key_to_id = {
-        batch_key: batch_id for batch_id, batch_key in batch_id_to_key.items()
-    }
-
-    def compute_sample_for_contest(
-        contest: Contest, sample_size: SampleSize
-    ) -> List[BallotDraw]:
-        # Compute the total number of ballot samples in all rounds leading up to
-        # this one. Note that this corresponds to the number of SampledBallotDraws,
-        # not SampledBallots.
-        num_previously_sampled = (
-            SampledBallotDraw.query.filter_by(contest_id=contest.id)
-            .join(SampledBallot)
-            .join(Batch)
-            .filter_by(has_cvrs=filter_has_cvrs)
-            .count()
-        )
-
-        # Create the pool of ballots to sample (aka manifest) by combining the
-        # manifests from every jurisdiction in the contest's universe.
-        manifest = {
-            batch_id_to_key[batch.id]: list(range(1, batch.num_ballots + 1))
-            for jurisdiction in contest.jurisdictions
-            for batch in jurisdiction.batches
-            if batch.has_cvrs == filter_has_cvrs
-        }
-
-        if filter_has_cvrs is None:
-            sample_size_num = sample_size["size"]
-        elif filter_has_cvrs:
-            sample_size_num = sample_size["sizeCvr"]
-        else:
-            sample_size_num = sample_size["sizeNonCvr"]
-
-        # Do the math! i.e. compute the actual sample
-        sample = sampler.draw_sample(
-            str(election.random_seed),
-            dict(manifest),
-            sample_size_num,
-            num_previously_sampled,
-            # In hybrid audits, sample without replacement for the non-CVR
-            # ballots, and with replacement for the CVR ballots.
-            # All other audit types sample with replacement.
-            with_replacement=(True if filter_has_cvrs is None else filter_has_cvrs),
-        )
-        return [
-            BallotDraw(
-                batch_id=batch_key_to_id[batch_key],
-                ballot_position=ballot_position,
-                contest_id=contest.id,
-                ticket_number=ticket_number,
-            )
-            for (ticket_number, (batch_key, ballot_position), _) in sample
-        ]
-
-    # Draw a sample for each contest
-    samples = [
-        sample
-        for contest, sample_size in contest_sample_sizes
-        for sample in compute_sample_for_contest(contest, sample_size)
-    ]
-    return samples
 
 
 @background_task
